refactor: update import path for dependency injection middleware
